+++ app/Http/Controllers/API/PhoneController.php
@@ -62,6 +62,8 @@ class PhoneController extends Controller
 
     public function massDestroy(MassDestroyPhoneRequest $request)
     {
+        abort_if(Gate::denies('phone_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         Phone::whereIn('id', request('ids'))->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
