+++ app/Http/Controllers/API/StorageDeviceController.php
@@ -64,6 +64,8 @@ class StorageDeviceController extends Controller
 
     public function massDestroy(MassDestroyStorageDeviceRequest $request)
     {
+        abort_if(Gate::denies('storage_device_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         StorageDevice::whereIn('id', request('ids'))->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
