+++ app/Http/Controllers/API/VlanController.php
@@ -75,6 +75,8 @@ class VlanController extends Controller
 
     public function massDestroy(MassDestroyVlanRequest $request)
     {
+        abort_if(Gate::denies('vlan_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         Vlan::whereIn('id', request('ids'))->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
