+++ app/Http/Controllers/API/WorkstationController.php
@@ -58,6 +58,8 @@ class WorkstationController extends Controller
 
     public function massDestroy(MassDestroyWorkstationRequest $request)
     {
+        abort_if(Gate::denies('workstation_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         Workstation::whereIn('id', request('ids'))->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
